Get persistent grid working again.

+++ Shared/Noise.h
@@ -13,7 +13,7 @@ typedef struct {
 } Gerstner;
 
 float hash(float2 p);
-float4 fbmd_7(float3 x, float f, float a, float l, float p, int o);
+float4 fbmd_7(float3 x, float f, float a, float l, float p, float o);
 float4 fbm(float3 x, int octaves);
 Gerstner gerstner(float3 x, float r, float t);
 float3 gNoised2(float2 p);
